style: move viewAllScreen.js to screens folder
